Removed demonic reference in menu.c quit prompt
removal of an uncomfortable mention of a demon that a user should not be provoked with, no actual change to any functional code; also removed the other useless and annoying statements in that quit prompt, replaced with a simple question.

To aid those who may or may not also have to experience that quit prompt. Thanks. Please approve.